Changed project structure a bit, got images to show
